Create new test_* files, force failure when data does not exist.
