+++ tests/test_Arizona_2020-Primary.py
@@ -8,7 +8,7 @@ def test_presidential(dbname):
 
 # TODO get AZ munger to read write-in votes; then will have to add 451 for Bo 'Heir Archy' Garcia
 def test_statewide_totals(dbname):
-    assert(not e.data_exists("2020 Primary","Arizona",dbname=dbname) or e.contest_total(
+    assert( e.contest_total(
             "2020 Primary",
             "Arizona",
             "US Senate AZ (Democratic Party)",
@@ -18,7 +18,7 @@ def test_statewide_totals(dbname):
     )
 
 def test_state_senate_totals(dbname):
-    assert (not e.data_exists("2020 Primary","Arizona",dbname=dbname) or e.contest_total(
+    assert ( e.contest_total(
             "2020 Primary",
             "Arizona",
             "AZ Senate District 10 (Republican Party)",
@@ -28,7 +28,7 @@ def test_state_senate_totals(dbname):
     )
 
 def test_state_rep_totals(dbname):
-    assert (not e.data_exists("2020 Primary","Arizona",dbname=dbname) or e.contest_total(
+    assert ( e.contest_total(
             "2020 Primary",
             "Arizona",
             "AZ House District 6 (Democratic Party)",
@@ -39,7 +39,7 @@ def test_state_rep_totals(dbname):
 
 
 def test_congressional_totals(dbname):
-    assert (not e.data_exists("2020 Primary","Arizona",dbname=dbname) or e.contest_total(
+    assert ( e.contest_total(
             "2020 Primary",
             "Arizona",
             "US House AZ District 6 (Democratic Party)",
