+++ tests/test_Michigan_2018-General.py
@@ -1,0 +1,34 @@
+import election_data_analysis as e
+
+def test_mi_statewide_totals_18(dbname):
+    assert (not e.data_exists("2018 General","Michigan",dbname=dbname) or e.contest_total(
+            "2018 General",
+            "Michigan",
+            "MI Governor",
+            dbname=dbname,
+        )
+            == 4250585
+    )
+
+
+def test_mi_senate_totals_18(dbname):
+    assert (not e.data_exists("2018 General","Michigan",dbname=dbname) or e.contest_total(
+            "2018 General",
+            "Michigan",
+            "MI Senate District 37",
+            dbname=dbname,
+        )
+            == 124414
+    )
+
+def test_mi_house_totals_18(dbname):
+    assert (not e.data_exists("2018 General","Michigan",dbname=dbname) or e.contest_total(
+            "2018 General",
+            "Michigan",
+            "MI House District 8",
+            dbname=dbname,
+        )
+            == 28017
+    )
+
+# vote types not available
