+++ tests/test_dataloader.py
@@ -1,278 +1,4 @@
 import election_data_analysis as e
-import pytest
-
-
-
-#NC18 tests
-
-#NC20 Tests
-
-def test_nc_presidential_20(dbname):
-    assert (not e.data_exists("2020 Primary","North Carolina",dbname=dbname) or e.contest_total(
-            "2020 Primary",
-            "North Carolina",
-            "US President (NC) (Democratic Party)",
-            dbname=dbname,
-        )
-            == 1331366
-    )
-
-
-def test_nc_statewide_totals_20(dbname):
-    assert (not e.data_exists("2020 Primary","North Carolina",dbname=dbname) or e.contest_total(
-            "2020 Primary",
-            "North Carolina",
-            "NC Governor (Democratic Party)",
-            dbname=dbname,
-        )
-            == 1293652
-    )
-
-
-def test_nc_senate_totals_20(dbname):
-    assert (not e.data_exists("2020 Primary","North Carolina",dbname=dbname) or e.contest_total(
-            "2020 Primary",
-            "North Carolina",
-            "US Senate NC (Democratic Party)",
-            dbname=dbname,
-        )
-            == 1260090
-    )
-
-
-def test_nc_rep_20_20(dbname):
-    assert (not e.data_exists("2020 Primary","North Carolina",dbname=dbname) or e.contest_total(
-            "2020 Primary",
-            "North Carolina",
-            "US House NC District 4 (Republican Party)",
-            dbname=dbname,
-        )
-            == 36096
-    )
-
-
-def test_nc_contest_by_vote_type_20(dbname):
-    assert ( not e.data_exists("2020 Primary","North Carolina", dbname=dbname) or
-            e.count_type_total(
-            "2016 General",
-            "North Carolina",
-"US House NC District 4 (Republican Party)",
-            "absentee-mail",
-            dbname=dbname,
-        )
-            == 426
-    )
-
-
-def test_nc_totals_match_vote_type_20(dbname):
-    assert (not e.data_exists("2020 General","North Carolina", dbname=dbname) or 
-            e.check_totals_match_vote_types("2020 General","North Carolina" ,dbname=dbname) == True)
-
-
-
-### Florida Data Loading Tests ###
-#FL16 test
-
-def test_fl_presidential(dbname):
-    assert (not e.data_exists("2016 General","Florida",dbname=dbname) or e.contest_total(
-            "2016 General",
-            "Florida",
-            "US President (FL)",
-            dbname=dbname,
-        )
-            == 9420039
-    )
-
-
-def test_fl_statewide_totals(dbname):
-    assert (not e.data_exists("2016 General","Florida",dbname=dbname) or e.contest_total(
-            "2016 General",
-            "Florida",
-            "US Senate FL",
-            dbname=dbname,
-        )
-            == 9301820
-    )
-
-
-def test_fl_senate_totals(dbname):
-    assert (not e.data_exists("2016 General","Florida",dbname=dbname) or e.contest_total(
-            "2016 General",
-            "Florida",
-            "FL Senate District 3",
-            dbname=dbname,
-        )
-            == 236480
-    )
-
-
-def test_fl_house_totals(dbname):
-    assert (not e.data_exists("2016 General","Florida",dbname=dbname) or e.contest_total(
-            "2016 General",
-            "Florida",
-            "US House FL District 10",
-            dbname=dbname,
-        )
-            == 305989
-    )
-
-
-def test_fl_contest_by_vote_type(dbname):
-    # Vote type not available
-    assert True == True
-
-
-def test_fl_totals_match_vote_type(dbname):
-    # Vote type not available
-    assert True == True
-
-#FL18 test
-
-def test_fl_presidential_18(dbname):
-    assert True == True
-
-
-def test_fl_statewide_totals_18(dbname):
-    assert (not e.data_exists("2018 General","Florida",dbname=dbname) or e.contest_total(
-            "2018 General",
-            "Florida",
-            "US Senate FL",
-            dbname=dbname,
-        )
-            == 8190005
-    )
-
-
-def test_fl_senate_totals_18(dbname):
-    assert (not e.data_exists("2018 General","Florida",dbname=dbname) or e.contest_total(
-            "2018 General",
-            "Florida",
-            "FL Senate District 4",
-            dbname=dbname,
-        )
-            == 235459
-    )
-
-
-def test_fl_house_totals_18(dbname):
-    assert (not e.data_exists("2018 General","Florida",dbname=dbname) or e.contest_total(
-            "2018 General",
-            "Florida",
-            "FL House District 11",
-            dbname=dbname,
-        )
-            == 85479
-    )
-
-
-def test_fl_contest_by_vote_type_18(dbname):
-    # Vote type not available
-    assert True == True
-
-
-def test_fl_totals_match_vote_type_18(dbname):
-    # Vote type not available
-    assert True == True
-
-#FL20 test
-
-def test_fl_presidential_20(dbname):
-    assert (not e.data_exists("2020 Primary","Florida",dbname=dbname) or e.contest_total(
-            "2020 Primary",
-            "Florida",
-            "US President (FL)",
-            dbname=dbname,
-        )
-            == 5958306
-    )
-
-
-def test_fl_statewide_totals_20(dbname):
-    assert True == True
-
-
-def test_fl_senate_totals_20(dbname):
-    assert (not e.data_exists("2020 Primary","Florida",dbname=dbname) or e.contest_total(
-            "2020 Primary",
-            "Florida",
-            "FL Senate District 21",
-            dbname=dbname,
-        )
-            == 54988
-    )
-
-
-def test_fl_house_totals_20(dbname):
-    assert (not e.data_exists("2020 Primary","Florida",dbname=dbname) or e.contest_total(
-            "2020 Primary",
-            "Florida",
-            "FL House District 11",
-            dbname=dbname,
-        )
-            == 17445
-    )
-
-
-def test_fl_contest_by_vote_type_20(dbname):
-    # Vote type not available
-    assert True == True
-
-
-def test_fl_totals_match_vote_type_20(dbname):
-    # Vote type not available
-    assert True == True
-
-### Pennsylvania Data Loading Tests ###
-
-#PA18 test
-
-
-#PA20
-
-def test_pa_presidential_20(dbname):
-    assert (not e.data_exists("2020 General","Pennsylvania",dbname=dbname) or e.contest_total(
-            "2020 General",
-            "Pennsylvania",
-            "US President (PA)",
-            dbname=dbname,
-        )
-            == 2739007
-    )
-
-
-def test_pa_statewide_totals_20(dbname):
-    assert (not e.data_exists("2020 General","Pennsylvania",dbname=dbname) or e.contest_total(
-            "2020 General",
-            "Pennsylvania",
-            "PA Governor",
-            dbname=dbname,
-        )
-            == 2484582
-    )
-
-
-def test_pa_senate_totals_20(dbname):
-    assert (not e.data_exists("2020 General","Pennsylvania",dbname=dbname) or e.contest_total(
-            "2020 General",
-            "Pennsylvania",
-            "PA Senate District 20",
-            dbname=dbname,
-        )
-            == 67898
-    )
-
-
-def test_pa_house_totals_20(dbname):
-    assert (not e.data_exists("2020 General","Pennsylvania",dbname=dbname) or e.contest_total(
-            "2020 General",
-            "Pennsylvania",
-            "PA House District 100",
-            dbname=dbname,
-        )
-            == 6327
-    )
-
-
 
 ### Arkansas Data Loading Tests ###
 #AR18 test
@@ -282,217 +8,6 @@ def test_ar_presidential_18(dbname):
     assert True == True
 
 
-def test_ar_statewide_totals_18(dbname):
-    assert (not e.data_exists("2018 General","Arkansas",dbname=dbname) or e.contest_total(
-            "2018 General",
-            "Arkansas",
-            "AR Governor",
-            dbname=dbname,
-        )
-            == 891509
-    )
-
-
-def test_ar_senate_totals_18(dbname):
-    assert (not e.data_exists("2018 General","Arkansas",dbname=dbname) or e.contest_total(
-            "2018 General",
-            "Arkansas",
-            "AR Senate District 5",
-            dbname=dbname,
-        )
-            == 27047
-    )
-
-
-def test_ar_house_totals_18(dbname):
-    assert (not e.data_exists("2018 General","Arkansas",dbname=dbname) or e.contest_total(
-            "2018 General",
-            "Arkansas",
-            "AR House District 19",
-            dbname=dbname,
-        )
-            == 7927
-    )
-
-
-def test_ar_contest_by_vote_type_18(dbname):
-    assert ( not e.data_exists("2018 General","Arkansas", dbname=dbname) or
-            e.count_type_total(
-            "2016 General",
-            "North Carolina",
-"AR Senate District 5",
-            "absentee",
-            dbname=dbname,
-        )
-            == 453
-    )
-
-
-def test_ar_totals_match_vote_type_18(dbname):
-    assert (not e.data_exists("2018 General","Arkansas", dbname=dbname) or 
-            e.check_totals_match_vote_types("2018 General","Arkansas" ,dbname=dbname) == True)
-
-#AR18 test
-
-def test_ar_presidential_20(dbname):
-    assert (not e.data_exists("2020 Primary","Arkansas",dbname=dbname) or e.contest_total(
-            "2020 Primary",
-            "Arkansas",
-            "US President (AR) (Republican Party)",
-            dbname=dbname,
-        )
-            == 246037
-    )
-
-
-def test_ar_statewide_totals_20(dbname):
-    assert (not e.data_exists("2020 Primary","Arkansas",dbname=dbname) or e.contest_total(
-            "2020 Primary",
-            "Arkansas",
-            "AR Governor",
-            dbname=dbname,
-        )
-            == 891509
-    )
-
-
-def test_ar_senate_totals_20(dbname):
-    assert (not e.data_exists("2020 Primary","Arkansas",dbname=dbname) or e.contest_total(
-            "2020 Primary",
-            "Arkansas",
-            "AR Senate District 5",
-            dbname=dbname,
-        )
-            == 27047
-    )
-
-
-def test_ar_house_totals_20(dbname):
-    assert (not e.data_exists("2020 Primary","Arkansas",dbname=dbname) or e.contest_total(
-            "2020 Primary",
-            "Arkansas",
-            "AR House District 19",
-            dbname=dbname,
-        )
-            == 7927
-    )
-
-
-def test_ar_contest_by_vote_type_20(dbname):
-    assert ( not e.data_exists("2020 Primary","Arkansas", dbname=dbname) or
-            e.count_type_total(
-            "2016 General",
-            "North Carolina",
-"AR Senate District 5",
-            "absentee",
-            dbname=dbname,
-        )
-            == 453
-    )
-
-
-def test_ar_totals_match_vote_type_20(dbname):
-    assert (not e.data_exists("2020 Primary","Arkansas", dbname=dbname) or 
-            e.check_totals_match_vote_types("2020 Primary","Arkansas" ,dbname=dbname) == True)
-
-
-### Michigan Data Loading Tests ###
-#MI16 test
-
-def test_mi_presidential_16(dbname):
-    assert (not e.data_exists("2016 General","Michigan",dbname=dbname) or e.contest_total(
-            "2016 General",
-            "Michigan",
-            "US President (MI)",
-            dbname=dbname,
-        )
-            == 4799284
-    )
-
-
-def test_mi_statewide_totals_16(dbname):
-    assert True == True
-
-
-def test_mi_us_rep_totals_16(dbname):
-    assert (not e.data_exists("2016 General","Michigan",dbname=dbname) or e.contest_total(
-            "2016 General",
-            "Michigan",
-            "US House MI District 4",
-            dbname=dbname,
-        )
-            == 315751
-    )
-
-def test_mi_house_totals_16(dbname):
-    assert (not e.data_exists("2016 General","Michigan",dbname=dbname) or e.contest_total(
-            "2016 General",
-            "Michigan",
-            "MI House District 8",
-            dbname=dbname,
-        )
-            == 34742
-    )
-
-
-def test_mi_contest_by_vote_type_16(dbname):
-    # Vote type not available
-    assert True == True
-
-
-def test_mi_totals_match_vote_type_16(dbname):
-    # Vote type not available
-    assert True == True
-
-#MI18 test
-
-def test_mi_presidential_18(dbname):
-    #no presidential contests in 2018
-    assert True == True
-
-
-def test_mi_statewide_totals_18(dbname):
-    assert (not e.data_exists("2018 General","Michigan",dbname=dbname) or e.contest_total(
-            "2018 General",
-            "Michigan",
-            "MI Governor",
-            dbname=dbname,
-        )
-            == 4250585
-    )
-
-
-def test_mi_senate_totals_18(dbname):
-    assert (not e.data_exists("2018 General","Michigan",dbname=dbname) or e.contest_total(
-            "2018 General",
-            "Michigan",
-            "MI Senate District 37",
-            dbname=dbname,
-        )
-            == 124414
-    )
-
-def test_mi_house_totals_18(dbname):
-    assert (not e.data_exists("2018 General","Michigan",dbname=dbname) or e.contest_total(
-            "2018 General",
-            "Michigan",
-            "MI House District 8",
-            dbname=dbname,
-        )
-            == 28017
-    )
-
-
-def test_mi_contest_by_vote_type_18(dbname):
-    # Vote type not available
-    assert True == True
-
-
-def test_mi_totals_match_vote_type_18(dbname):
-    # Vote type not available
-    assert True == True
-
-#MI20 test
 
 def test_mi_statewide_totals_20(dbname):
     assert (not e.data_exists("2020 Primary","Michigan",dbname=dbname) or e.contest_total(
